# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "东软顾养中心管理系统API文档",
        "title": "东软顾养中心 API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8081",
    "basePath": "/api",
    "paths": {
...
        "/care-records/{id}/records": {
            "post": {
                "description": "向指定护理记录文档中追加一条详细记录",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "护理记录管理"
                ],
                "summary": "追加护理记录",
                "parameters": [
                    {
                        "type": "string",
                        "description": "记录ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "详细记录信息",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/domain.RecordItems"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "追加成功",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
        "/records": {
            "get": {
                "description": "分页获取所有登记记录",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "登记管理"
                ],
                "summary": "获取登记记录列表",
                "parameters": [
                    {
                        "type": "string",
                        "description": "客户ID",
                        "name": "customer_id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "记录类型",
                        "name": "type",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 0,
                        "description": "跳过数量",
                        "name": "skip",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 20,
                        "description": "每页数量",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "获取成功",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/records/check-in": {
            "post": {
                "description": "为客户办理入住登记，创建客户档案，自动分配床位并更新客户状态",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "登记管理"
                ],
                "summary": "入住登记",
                "parameters": [
                    {
                        "description": "入住信息",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/domain.ElderlyRegisterRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "入住登记成功",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/records/check-in-info": {
            "get": {
                "description": "根据条件筛选获取入住登记记录",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "登记管理"
                ],
                "summary": "获取入住列表",
                "parameters": [
                    {
                        "type": "string",
                        "description": "客户姓名",
                        "name": "name",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "房间号",
                        "name": "room_number",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "护理级别",
                        "name": "nursing_level",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "入住开始日期",
                        "name": "check_in_start",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "入住结束日期",
                        "name": "check_in_end",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "获取成功",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/records/check-out": {
            "post": {
                "description": "为客户办理退住登记，释放床位并更新客户状态",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "登记管理"
                ],
                "summary": "退住登记",
                "parameters": [
                    {
                        "description": "退住信息",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/domain.CheckOutRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "退住登记成功",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/records/check-out-list": {
            "get": {
                "description": "根据条件筛选获取退住登记记录",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "登记管理"
                ],
                "summary": "获取退住列表",
                "parameters": [
                    {
                        "type": "string",
                        "description": "客户姓名",
                        "name": "name",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "房间号",
                        "name": "room_number",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "退住原因",
                        "name": "reason",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "退住开始日期",
                        "name": "check_out_start",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "退住结束日期",
                        "name": "check_out_end",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "获取成功",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/records/customer/{customer_id}": {
            "get": {
                "description": "根据客户ID获取其所有登记记录",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "登记管理"
                ],
                "summary": "获取客户登记记录",
                "parameters": [
                    {
                        "type": "string",
                        "description": "客户ID",
                        "name": "customer_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "获取成功",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/records/outgoing": {
            "post": {
                "description": "办理客户外出登记",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "登记管理"
                ],
                "summary": "外出登记",
                "parameters": [
                    {
                        "description": "外出信息",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/domain.OutgoingRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "外出登记成功",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/records/outgoing-list": {
            "get": {
                "description": "根据条件筛选获取外出登记记录",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "登记管理"
                ],
                "summary": "获取外出列表",
                "parameters": [
                    {
                        "type": "string",
                        "description": "客户姓名",
                        "name": "name",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "开始日期",
                        "name": "start_date",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "结束日期",
                        "name": "end_date",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "状态",
                        "name": "status",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "获取成功",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/records/return": {
            "post": {
                "description": "办理客户外出返回登记",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "登记管理"
                ],
                "summary": "外出返回",
                "parameters": [
                    {
                        "description": "返回信息",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/domain.ReturnRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "返回登记成功",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "domain.CheckOutRequest": {
            "type": "object",
            "required": [
                "created_by",
                "customer_id"
            ],
            "properties": {
                "created_by": {
                    "type": "string"
                },
                "customer_id": {
                    "type": "string"
                },
                "note": {
                    "type": "string"
                }
            }
        },
...
        "domain.ElderlyRegisterRequest": {
            "type": "object",
            "required": [
                "age",
                "bed_id",
                "contact_name",
                "contact_phone",
                "dietary_type",
                "gender",
                "health_level",
                "id_card",
                "nursing_level",
                "relationship"
            ],
            "properties": {
                "age": {
                    "description": "年龄",
                    "type": "integer"
                },
                "allergy_history": {
                    "description": "过敏史",
                    "type": "string"
                },
                "bed_id": {
                    "description": "床位分配",
                    "type": "string"
                },
                "check_in_date": {
                    "description": "入住安排详情",
                    "type": "string"
                },
                "contact_address": {
                    "description": "联系地址",
                    "type": "string"
                },
                "contact_name": {
                    "description": "紧急联系人信息",
                    "type": "string"
                },
                "contact_phone": {
                    "description": "联系电话",
                    "type": "string"
                },
                "dietary_type": {
                    "description": "膳食类型",
                    "type": "string"
                },
                "gender": {
                    "description": "性别 (男/女)",
                    "type": "string"
                },
                "health_level": {
                    "description": "健康状况记录",
                    "type": "string"
                },
                "home_address": {
                    "description": "家庭住址",
                    "type": "string"
                },
                "id_card": {
                    "description": "身份证号",
                    "type": "string"
                },
                "medical_history": {
                    "description": "既往病史",
                    "type": "string"
                },
                "medication": {
                    "description": "用药情况",
                    "type": "string"
                },
                "name": {
                    "description": "基本信息",
                    "type": "string"
                },
                "nursing_level": {
                    "description": "护理级别",
                    "type": "string"
                },
                "phone_number": {
                    "description": "联系电话",
                    "type": "string"
                },
                "relationship": {
                    "description": "关系",
                    "type": "string"
                },
                "remarks": {
                    "description": "备注说明",
                    "type": "string"
                }
            }
        },
...
        "domain.OutgoingRequest": {
            "type": "object",
            "required": [
                "created_by",
                "customer_id"
            ],
            "properties": {
                "created_by": {
                    "type": "string"
                },
                "customer_id": {
                    "type": "string"
                },
                "note": {
                    "type": "string"
                }
            }
        },
...
        "domain.RecordItems": {
            "type": "object",
            "required": [
                "care_item",
                "care_personnel",
                "care_result",
                "care_time"
            ],
            "properties": {
                "care_item": {
                    "description": "护理项目",
                    "type": "string"
                },
                "care_personnel": {
                    "description": "护理人员",
                    "type": "string"
                },
                "care_result": {
                    "description": "护理结果",
                    "type": "string"
                },
                "care_time": {
                    "description": "护理时间",
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                }
            }
        },
        "domain.ReturnRequest": {
            "type": "object",
            "required": [
                "created_by",
                "customer_id"
            ],
            "properties": {
                "created_by": {
                    "type": "string"
                },
                "customer_id": {
                    "type": "string"
                },
                "note": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "Bearer": {
            "description": "Type \"Bearer\" followed by a space and JWT token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}